requirements.txt
cfscrape
progressbar2
